+++ doc/changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 2.2.8
+
+- Fixed: Magnebot crashes on version check.
+- Fixed: The Magnebot can't turn when no frames are skipped. If no frames are skipped, the Magnebot's wheels have slightly greater force limits.
+- Required version of TDW is now >=1.12.16.0
+
 ## 2.2.7
 
 - Fixed: Magnebot can't be used with Python 3.10 or newer.